Update format-numeric helper
